# docs/backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Renda+ ERP — API do servidor
  version: v1
  description: |
    Contrato da API usada pelo app do Mac (Sprints 1 a 4).
    Regras gerais:
    - Toda rota exige sessão (`Authorization: Bearer <token>`), exceto `GET /api/v1/status` e `POST /api/v1/session` (ADR-005).
      Sem sessão válida: 401 `UNAUTHENTICATED`. Perfil sem a permissão: 403 `FORBIDDEN` (a negativa fica na auditoria).
    - Edição exige `If-Match` com a versão lida (ETag); versão desatualizada: 412 `VERSION_MISMATCH`; sem cabeçalho: 428.
    - Cadastro exige `Idempotency-Key` (8 a 100 caracteres); repetir com a mesma chave devolve o mesmo registro;
      mesma chave com outro conteúdo: 422 `IDEMPOTENCY_KEY_REUSED`.
    - Cliente e fornecedor são papéis do mesmo parceiro, cada um com a sua situação. Cadastrar num papel o CNPJ de um parceiro
      que ainda não o tem devolve 422 `PARTNER_OTHER_ROLE`, com `details` `partnerId` e `version` para chamar `.../{id}/enable`.
    - Valores decimais (custo de referência, fator de conversão) trafegam como texto com ponto, até 6 casas (ADR-006, PD-007).
    - Valores em reais trafegam em centavos, como texto de inteiro (`"30207146"` = R$ 302.071,46); datas de negócio `AAAA-MM-DD`.
    - Documentos comerciais (Sprint 4): a revisão emitida da proposta é imutável; o pedido só muda em rascunho. Confirmar o pedido
      (`Idempotency-Key` + `If-Match`) cria o projeto, um equipamento por unidade das linhas de equipamento e um título a receber
...
servers:
...
security:
  - sessao: []
...
paths:
...
  /api/v1/sales-orders:
    get:
      tags: [Pedidos]
      summary: Lista de pedidos (sales_order.read)
      parameters:
        - { name: search, in: query, required: false, schema: { type: string }, description: 'Número, cliente ou unidade' }
        - { name: status, in: query, required: false, schema: { type: string, enum: [DRAFT, CONFIRMED, IN_EXECUTION, COMPLETED, CANCELLED, TODOS] } }
      responses:
        '200': { description: Pedidos, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/SalesOrderSummary' } } } } }
    post:
      tags: [Pedidos]
      summary: Pedido em rascunho (sales_order.create), idempotente; número PV00001
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SalesOrderRequest' } } }
      responses:
        '201': { description: Cadastrado, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/SalesOrder' } } } }
        '403': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
  /api/v1/sales-orders/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Pedidos]
      summary: Pedido com linhas, parcelas e, depois de confirmado, projeto, equipamentos e títulos (sales_order.read)
      responses:
        '200': { description: Pedido, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/SalesOrder' } } } }
        '404': { $ref: '#/components/responses/Erro' }
    put:
      tags: [Pedidos]
      summary: Altera o rascunho (sales_order.update); confirmado devolve 409 (mudança só por aditivo)
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SalesOrderRequest' } } }
      responses:
        '200': { description: Nova versão, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/SalesOrder' } } } }
        '409': { $ref: '#/components/responses/Erro' }
        '412': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
        '428': { $ref: '#/components/responses/Erro' }
  /api/v1/sales-orders/{id}/confirmations:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Pedidos]
      summary: >-
        Confirma — ConfirmSalesOrder (sales_order.confirm): projeto, equipamentos e um título por parcela numa transação, uma única
        vez. Σ parcelas ≠ total: 422 ORDER_INVALID; cliente inativo ou unidade de outro cliente: 422 PARTNER_INACTIVE_OR_UNIT_MISMATCH
      parameters: [{ $ref: '#/components/parameters/IfMatch' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '200': { description: Pedido confirmado (ou a confirmação existente), headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/SalesOrder' } } } }
        '403': { $ref: '#/components/responses/Erro' }
        '409': { $ref: '#/components/responses/Erro' }
        '412': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
  /api/v1/sales-orders/{id}/cancellations:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Pedidos]
      summary: >-
        Cancela com motivo — CancelSalesOrder (sales_order.cancel). Confirmado: cancela os títulos, encerra o projeto e cancela os
        equipamentos; com recebimento ou execução iniciada, 422 CANCELLATION_BLOCKED_BY_EFFECTS (premissa PD-003)
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: false
        content: { application/json: { schema: { $ref: '#/components/schemas/ReasonRequest' } } }
      responses:
        '200': { description: Pedido cancelado, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/SalesOrder' } } } }
        '409': { $ref: '#/components/responses/Erro' }
        '412': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
  /api/v1/sales-orders/{id}/history:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Pedidos]
      summary: Histórico do pedido (sales_order.read)
      responses:
        '200': { description: Histórico, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/HistoryEntry' } } } } }
...
  /api/v1/equipment:
    get:
      tags: [Projetos e equipamentos]
      summary: Lista de equipamentos (equipment.read); cancelados só com includeCancelled
      parameters:
        - { name: search, in: query, required: false, schema: { type: string }, description: 'Código, modelo, série, projeto ou cliente' }
        - { name: projectId, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: includeCancelled, in: query, required: false, schema: { type: boolean, default: false } }
      responses:
        '200': { description: Equipamentos, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Equipment' } } } } }
  /api/v1/equipment/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Projetos e equipamentos]
      summary: Ficha do equipamento (equipment.read)
      responses:
        '200': { description: Equipamento, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Equipment' } } } }
        '404': { $ref: '#/components/responses/Erro' }
    put:
      tags: [Projetos e equipamentos]
      summary: Altera série e observações — UpdateEquipment (equipment.update); série única por modelo (422 EQUIPMENT_SERIAL_DUPLICATE)
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/EquipmentRequest' } } }
      responses:
        '200': { description: Nova versão, headers: { ETag: { $ref: '#/components/headers/ETag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Equipment' } } } }
        '409': { $ref: '#/components/responses/Erro' }
        '412': { $ref: '#/components/responses/Erro' }
        '422': { $ref: '#/components/responses/Erro' }
        '428': { $ref: '#/components/responses/Erro' }
  /api/v1/equipment/{id}/history:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Projetos e equipamentos]
      summary: Histórico do equipamento (equipment.read)
      responses:
        '200': { description: Histórico, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/HistoryEntry' } } } } }
...
components:
  securitySchemes:
    sessao:
      type: http
      scheme: bearer
      description: Token opaco devolvido por POST /api/v1/session
  headers:
    ETag:
      description: Versão do registro entre aspas, para usar no If-Match
      schema: { type: string, example: '"3"' }
  parameters:
    Id:
      { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    IfMatch:
      { name: If-Match, in: header, required: true, schema: { type: string, example: '"3"' }, description: 'Versão lida (ETag)' }
    IdempotencyKey:
      { name: Idempotency-Key, in: header, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_-]{8,100}$' } }
  responses:
    Erro:
      description: Erro padronizado
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
  schemas:
    ApiError:
      type: object
      properties:
        code: { type: string, example: VERSION_MISMATCH }
        message: { type: string }
        details: { type: array, items: { type: object, properties: { field: { type: [string, 'null'] }, message: { type: string } } } }
        correlationId: { type: string }
        retryable: { type: boolean }
...
    HistoryEntry:
      type: object
      properties:
        occurredAt: { type: string, format: date-time }
        actor: { type: string }
        action:
          type: string
          description: >-
            PARTNER_REGISTERED, PARTNER_UPDATED, PARTNER_DEACTIVATED, PARTNER_ROLE_ENABLED, ITEM_REGISTERED, ITEM_UPDATED,
            ITEM_DEACTIVATED, PROPOSAL_DRAFTED, PROPOSAL_UPDATED, PROPOSAL_REVISION_ISSUED, PROPOSAL_REVISION_CREATED,
            PROPOSAL_LOST, PROPOSAL_WON, SALES_ORDER_DRAFTED, SALES_ORDER_UPDATED, SALES_ORDER_CONFIRMED, SALES_ORDER_CANCELLED,
            PROJECT_CREATED, PROJECT_CLOSED, EQUIPMENT_CREATED, EQUIPMENT_UPDATED, EQUIPMENT_CANCELLED
        version: { type: integer }
        reason: { type: [string, 'null'] }
        changes:
          type: object
          additionalProperties:
            type: object
            properties:
              before: { type: [string, 'null'] }
              after: { type: [string, 'null'] }
...
    Cents:
      type: string
      pattern: '^\d{1,18}$'
      description: Valor em centavos, texto de inteiro ("30207146" = R$ 302.071,46)
    SalesLineRequest:
      type: object
      required: [kind, quantity, unitPrice]
      properties:
        id: { type: [string, 'null'], description: 'Id da linha existente; vazio numa linha nova' }
        kind: { type: string, enum: [EQUIPAMENTO, MATERIAL, SERVICO] }
        itemId: { type: [string, 'null'], format: uuid, description: 'Obrigatório para MATERIAL e SERVICO (mesma natureza); vazio para EQUIPAMENTO' }
        description: { type: [string, 'null'], maxLength: 200, description: 'Modelo do equipamento; nos itens, a descrição do cadastro quando vazio' }
        quantity: { type: string, example: '10.5', description: 'Até 6 casas; equipamento em unidades inteiras (1 a 100)' }
        unitPrice: { type: string, example: '16.33', description: 'Preço unitário em reais, até 6 casas' }
        discountCents: { type: [string, 'null'], example: '10000' }
    SalesLine:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [EQUIPAMENTO, MATERIAL, SERVICO] }
        itemId: { type: [string, 'null'], format: uuid }
        itemCode: { type: [string, 'null'] }
        description: { type: string }
        quantity: { type: string }
        uom: { type: string }
        unitPrice: { type: string }
        discountCents: { $ref: '#/components/schemas/Cents' }
        grossCents: { $ref: '#/components/schemas/Cents' }
        totalCents: { $ref: '#/components/schemas/Cents', description: 'round(quantidade × preço) − desconto, meio-par por linha (PD-002)' }
...
    ReasonRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, maxLength: 500 }
    Installment:
      type: object
      required: [dueDate, amountCents]
      properties:
        seq: { type: [integer, 'null'], description: 'Na resposta: sequência a partir de 1 (a ordem da lista)' }
        dueDate: { type: string, format: date }
        amountCents: { $ref: '#/components/schemas/Cents' }
        milestone: { type: [string, 'null'], maxLength: 200 }
    SalesOrderRequest:
      type: object
      required: [customerId, unitId, contractDate, lines]
      properties:
        customerId: { type: string, format: uuid }
        unitId: { type: string, format: uuid }
        contractDate: { type: string, format: date }
        promisedDate: { type: [string, 'null'], format: date }
        notes: { type: [string, 'null'], maxLength: 1000 }
        lines: { type: array, items: { $ref: '#/components/schemas/SalesLineRequest' } }
        installments: { type: array, items: { $ref: '#/components/schemas/Installment' }, description: 'Na confirmação, a soma precisa ser o total exato' }
    SalesOrderSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        customerCode: { type: string }
        customerName: { type: string }
        unitName: { type: string }
        proposalCode: { type: [string, 'null'] }
        contractDate: { type: string, format: date }
        totalCents: { $ref: '#/components/schemas/Cents' }
        status: { type: string, enum: [DRAFT, CONFIRMED, IN_EXECUTION, COMPLETED, CANCELLED] }
        version: { type: string }
    SalesOrder:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: PV00001 }
        customerId: { type: string, format: uuid }
        customerCode: { type: string }
        customerName: { type: string }
        unitId: { type: string, format: uuid }
        unitName: { type: string }
        proposalId: { type: [string, 'null'], format: uuid }
        proposalCode: { type: [string, 'null'] }
        proposalRevision: { type: [integer, 'null'] }
        contractDate: { type: string, format: date }
        promisedDate: { type: [string, 'null'], format: date }
        notes: { type: [string, 'null'] }
        status: { type: string, enum: [DRAFT, CONFIRMED, IN_EXECUTION, COMPLETED, CANCELLED] }
        totalCents: { $ref: '#/components/schemas/Cents' }
        scheduledCents: { $ref: '#/components/schemas/Cents' }
        lines: { type: array, items: { $ref: '#/components/schemas/SalesLine' } }
        installments: { type: array, items: { $ref: '#/components/schemas/Installment' } }
        confirmedAt: { type: [string, 'null'], format: date-time }
        confirmedBy: { type: [string, 'null'] }
        projectId: { type: [string, 'null'], format: uuid }
        projectCode: { type: [string, 'null'] }
        projectStage: { type: [string, 'null'] }
        equipment:
          type: array
          items: { type: object, properties: { id: { type: string }, code: { type: string }, model: { type: string }, serialNumber: { type: [string, 'null'] }, status: { type: string } } }
        titles:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              code: { type: string }
              label: { type: string }
              dueDate: { type: string, format: date }
              competence: { type: string, example: '2026-10' }
              originalCents: { $ref: '#/components/schemas/Cents' }
              balanceCents: { $ref: '#/components/schemas/Cents' }
              status: { type: string, enum: [OPEN, PARTIAL, SETTLED, RENEGOTIATED, CANCELLED] }
        cancelledAt: { type: [string, 'null'], format: date-time }
        cancelledBy: { type: [string, 'null'] }
        cancelReason: { type: [string, 'null'] }
        version: { type: string }
        createdAt: { type: string, format: date-time }
        createdBy: { type: string }
        updatedAt: { type: [string, 'null'], format: date-time }
        updatedBy: { type: [string, 'null'] }
...
    EquipmentRequest:
      type: object
      properties:
        serialNumber: { type: [string, 'null'], maxLength: 60 }
        notes: { type: [string, 'null'], maxLength: 1000 }
    Equipment:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: EQ00001 }
        model: { type: string }
        itemId: { type: [string, 'null'], format: uuid }
        projectId: { type: string, format: uuid }
        projectCode: { type: string }
        orderCode: { type: string }
        customerId: { type: string, format: uuid }
        customerCode: { type: string }
        customerName: { type: string }
        unitId: { type: string, format: uuid }
        unitName: { type: string }
        serialNumber: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        status: { type: string, enum: [ATIVO, CANCELADO] }
        acceptedOn: { type: [string, 'null'], format: date, description: 'Do aceite registrado (PD-015); vazio até a instalação' }
        warrantyStart: { type: [string, 'null'], format: date }
        version: { type: string }
        createdAt: { type: string, format: date-time }
        createdBy: { type: string }
        updatedAt: { type: [string, 'null'], format: date-time }
        updatedBy: { type: [string, 'null'] }
...